[arcane] Déplace 'ComponentItemVector' et les classes dérivées dans 'arcane/core/materials'.
