+++ arcane/src/arcane/core/materials/MaterialsCoreGlobal.h
@@ -59,9 +59,24 @@ class ComponentItemVectorView;
 class ComponentPartItemVectorView;
 class ComponentPurePartItemVectorView;
 class ComponentImpurePartItemVectorView;
+class MeshComponentPartData;
+class MatCellEnumerator;
+class ComponentItemVector;
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
+
+//! Liste de ComponentCell issues d'un IMeshComponent.
+typedef ComponentItemVector ComponentCellVector;
+
+//! Type de la vue sur un EnvCellVector
+typedef EnvItemVectorView EnvCellVectorView;
+
+//! Type de la vue sur un MatCellVector
+typedef MatItemVectorView MatCellVectorView;
+
+//! Type de la vue sur un ComponentCellVector
+typedef ComponentItemVectorView ComponentCellVectorView;
 
 //! Liste de composants multi-matériaux du maillage.
 typedef ConstArrayView<IMeshComponent*> MeshComponentList;
